updated ssr vs. csr mix for page
