+++ src/app/(proper_react)/(redesign)/(public)/how-it-works/components/FooterSection.tsx
@@ -2,13 +2,15 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { Props } from "../HowItWorksView";
 import { Button } from "../../../../../components/client/Button";
 import styles from "../HowItWorksView.module.scss";
+import { ExtendedReactLocalization } from "../../../../../functions/l10n";
 
-export const FooterSection = (props: Props) => {
-  const { l10n } = props;
-
+export const FooterSection = ({
+  l10n,
+}: {
+  l10n: ExtendedReactLocalization;
+}) => {
   return (
     <footer className={styles.footerSection}>
       <h2 className={styles.footerSectionCTA}>
